lastv/ophthalmic_care_system/requirements.txt
Flask==2.3.3 
Werkzeug==2.3.7 
tensorflow==2.13.0 
numpy==1.24.3 
opencv-python==4.8.0.76 
Pillow==10.0.0 
scipy==1.11.2 
scikit-learn==1.3.0 
matplotlib==3.7.2 
seaborn==0.12.2 
SQLAlchemy==2.0.20 
python-dotenv==1.0.0
